fix(state): :bug: fix recreating tabs not reacting to appState changes
the NoteSyncService was initialized with a dispatchers array which was created at initialization time, meaning that any changes to this array, e.g. when a tab would be closed and possibly reopened again, would not be taken into consideration. The consequence was that a tab would only work once, when the app was opened.