fix(state): :bug: fix chat messages not updating closing / reopening note view
The events watcher service was taking a dispatcher as a constructor parameter. When renewing the panel, the events watcher would still use the old dispatcher, therefore not updating the current comments view. With this new design, the watching of events is local to the component, which also means that we need a way to stop listening to events when we unmount the component. We did that by using the `return` method of the AsyncGenerator watch cursor.